Update secret name and move `getZcapClient()` to helpers file.
