+++ .github/workflows/main.yml
@@ -46,7 +46,7 @@ jobs:
         run: npm run generate-vcs
       - name: Run test with Node.js ${{ matrix.node-version }}
         env:
-          CLIENT_SECRET: ${{ secrets.CLIENT_SECRET }}
+          CLIENT_SECRET_DB: ${{ secrets.CLIENT_SECRET_DB }}
         run: npm run test
         continue-on-error: true
       - name: Deploy to Github Pages
