+++ tests/helpers.js
@@ -2,11 +2,16 @@
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
-const https = require('https');
-const {httpClient} = require('@digitalbazaar/http-client');
+const {decodeSecretKeySeed} = require('bnid');
 const {decodeList} = require('@digitalbazaar/vc-status-list');
+const didKey = require('@digitalbazaar/did-method-key');
+const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
+const {httpClient} = require('@digitalbazaar/http-client');
+const https = require('https');
+const {ZcapClient} = require('@digitalbazaar/ezcap');
 
 const agent = new https.Agent({rejectUnauthorized: false});
+const didKeyDriver = didKey.driver();
 
 // Javascript's default ISO timestamp is contains milliseconds.
 // This lops off the MS part of the UTC RFC3339 TimeStamp and replaces
@@ -51,9 +56,26 @@ const getCredentialStatus = async ({verifiableCredential}) => {
   return {status, statusListCredential: revocationListCredential};
 };
 
+async function getZcapClient() {
+  if(!process.env.CLIENT_SECRET_DB) {
+    throw new Error('ENV variable CLIENT_SECRET_DB is required.');
+  }
+  const secretKeySeed = process.env.CLIENT_SECRET_DB;
+  const seed = await decodeSecretKeySeed({secretKeySeed});
+  const didKey = await didKeyDriver.generate({seed});
+  const {didDocument: {capabilityInvocation}} = didKey;
+  const zcapClient = new ZcapClient({
+    SuiteClass: Ed25519Signature2020,
+    invocationSigner: didKey.keyPairs.get(capabilityInvocation[0]).signer(),
+    agent
+  });
+  return zcapClient;
+}
+
 module.exports = {
   ISOTimeStamp,
   deepClone,
   unwrapResponse,
-  getCredentialStatus
+  getCredentialStatus,
+  getZcapClient
 };
